Connected admin products page to the database and implemented the update feature
Changed the price from float to int

+++ E-Mart/src/java/Model/Camera.java
@@ -14,7 +14,7 @@ public class Camera extends BaseProduct {
 
     private String formFactor;
 
-    public Camera(int productId, String productName, float price, int quantity, String photos,
+    public Camera(int productId, String productName, int price, int quantity, String photos,
             String brand, String modelName, String productDescription, String category,
             String formFactor) {
         super(productId, productName, price, quantity, photos, brand, modelName, productDescription, category);
